+++ src/print-xml.lisp
@@ -41,6 +41,14 @@
     (format stream "~a</~a>~%"
             pad (name element))))
 
+(defmethod print-xml (stream (element <dict-section>) &optional (indent 0))
+  (let ((pad (make-indent indent)))
+    (format stream "~a<~a label=~S>~%"
+            pad (name element) (title element))
+    (print-children stream element indent)
+    (format stream "~a</~a>~%"
+            pad (name element))))
+
 (defmethod print-xml (stream (element <chapter>) &optional (indent 0))
   (let ((pad (make-indent indent)))
     (format stream "~a<chapter number=~S title=~S ref=~S ref-title=~S>~%"